added wagtail hooks
